chore: extract NpmPackage from NodeProject
In an effort to componentize the monolithic NodeProject so that the various aspects
of it can be reused by other projects through composition, this commit extracts the
management of `package.json` into `NpmPackage`.
