+++ .gitignore
@@ -47,8 +47,8 @@ tsconfig.json
 !/.projen/deps.json
 !/.gitpod.yml
 !/package.json
-!/.npmignore
 !/LICENSE
+!/.npmignore
 !/.projenrc.js
 !version.json
 !/.versionrc.json
