+++ package.json
@@ -9,86 +9,35 @@
     "projen": "bin/projen"
   },
   "scripts": {
-    "projen": "/bin/bash ./projen.bash",
-    "start": "/bin/bash ./projen.bash start",
-    "clobber": "/bin/bash ./projen.bash clobber",
-    "compile": "/bin/bash ./projen.bash compile",
-    "test:compile": "/bin/bash ./projen.bash test:compile",
-    "test": "/bin/bash ./projen.bash test",
     "build": "/bin/bash ./projen.bash build",
     "bump": "/bin/bash ./projen.bash bump",
-    "release": "/bin/bash ./projen.bash release",
-    "test:watch": "/bin/bash ./projen.bash test:watch",
-    "test:update": "/bin/bash ./projen.bash test:update",
-    "projen:upgrade": "/bin/bash ./projen.bash projen:upgrade",
-    "watch": "/bin/bash ./projen.bash watch",
-    "package": "/bin/bash ./projen.bash package",
-    "eslint": "/bin/bash ./projen.bash eslint",
+    "clobber": "/bin/bash ./projen.bash clobber",
     "compat": "/bin/bash ./projen.bash compat",
+    "compile": "/bin/bash ./projen.bash compile",
     "docgen": "/bin/bash ./projen.bash docgen",
-    "readme-macros": "/bin/bash ./projen.bash readme-macros"
+    "eslint": "/bin/bash ./projen.bash eslint",
+    "package": "/bin/bash ./projen.bash package",
+    "projen": "/bin/bash ./projen.bash",
+    "projen:upgrade": "/bin/bash ./projen.bash projen:upgrade",
+    "readme-macros": "/bin/bash ./projen.bash readme-macros",
+    "release": "/bin/bash ./projen.bash release",
+    "start": "/bin/bash ./projen.bash start",
+    "test": "/bin/bash ./projen.bash test",
+    "test:compile": "/bin/bash ./projen.bash test:compile",
+    "test:update": "/bin/bash ./projen.bash test:update",
+    "test:watch": "/bin/bash ./projen.bash test:watch",
+    "watch": "/bin/bash ./projen.bash watch"
   },
   "author": {
     "name": "Elad Ben-Israel",
     "email": "benisrae@amazon.com",
     "organization": false
   },
-  "devDependencies": {
-    "@types/fs-extra": "^8",
-    "@types/glob": "^7.1.3",
-    "@types/inquirer": "^7.3.1",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^10.17.0",
-    "@types/semver": "^7.3.4",
-    "@types/yargs": "^15.0.10",
-    "@typescript-eslint/eslint-plugin": "^4.3.0",
-    "@typescript-eslint/parser": "^4.3.0",
-    "eslint": "^7.14.0",
-    "eslint-import-resolver-node": "^0.3.4",
-    "eslint-import-resolver-typescript": "^2.3.0",
-    "eslint-plugin-import": "^2.22.1",
-    "jest": "^26.6.3",
-    "jest-junit": "^12",
-    "jsii": "^1.15.0",
-    "jsii-diff": "^1.15.0",
-    "jsii-docgen": "^1.5.2",
-    "jsii-pacmak": "^1.15.0",
-    "jsii-release": "^0.2.3",
-    "json-schema": "^0.2.5",
-    "markmac": "^0.1.23",
-    "standard-version": "^9.0.0",
-    "ts-jest": "^26.4.4",
-    "typescript": "^3.9.5"
-  },
-  "peerDependencies": {},
-  "dependencies": {
-    "@iarna/toml": "^2.2.5",
-    "chalk": "^4.1.0",
-    "decamelize": "^4.0.0",
-    "fs-extra": "^9.0.1",
-    "glob": "^7",
-    "inquirer": "^7.3.3",
-    "semver": "^7.3.2",
-    "yaml": "^1.10.0",
-    "yargs": "^16.1.1"
-  },
-  "bundledDependencies": [
-    "@iarna/toml",
-    "chalk",
-    "decamelize",
-    "fs-extra",
-    "glob",
-    "inquirer",
-    "semver",
-    "yaml",
-    "yargs"
-  ],
-  "keywords": [],
   "engines": {
     "node": ">= 10.17.0"
   },
-  "main": "lib/index.js",
   "license": "Apache-2.0",
+  "main": "lib/index.js",
   "version": "0.10.1",
   "jest": {
     "clearMocks": true,
@@ -130,5 +79,54 @@
       "rootDir": "src"
     }
   },
+  "devDependencies": {
+    "@types/fs-extra": "^8",
+    "@types/glob": "^7.1.3",
+    "@types/inquirer": "^7.3.1",
+    "@types/jest": "^26.0.15",
+    "@types/node": "^10.17.0",
+    "@types/semver": "^7.3.4",
+    "@types/yargs": "^15.0.10",
+    "@typescript-eslint/eslint-plugin": "^4.3.0",
+    "@typescript-eslint/parser": "^4.3.0",
+    "eslint": "^7.14.0",
+    "eslint-import-resolver-node": "^0.3.4",
+    "eslint-import-resolver-typescript": "^2.3.0",
+    "eslint-plugin-import": "^2.22.1",
+    "jest": "^26.6.3",
+    "jest-junit": "^12",
+    "jsii": "^1.15.0",
+    "jsii-diff": "^1.15.0",
+    "jsii-docgen": "^1.5.2",
+    "jsii-pacmak": "^1.15.0",
+    "jsii-release": "^0.2.3",
+    "json-schema": "^0.2.5",
+    "markmac": "^0.1.23",
+    "standard-version": "^9.0.0",
+    "ts-jest": "^26.4.4",
+    "typescript": "^3.9.5"
+  },
+  "dependencies": {
+    "@iarna/toml": "^2.2.5",
+    "chalk": "^4.1.0",
+    "decamelize": "^4.0.0",
+    "fs-extra": "^9.0.1",
+    "glob": "^7",
+    "inquirer": "^7.3.3",
+    "semver": "^7.3.2",
+    "yaml": "^1.10.0",
+    "yargs": "^16.1.1"
+  },
+  "bundledDependencies": [
+    "@iarna/toml",
+    "chalk",
+    "decamelize",
+    "fs-extra",
+    "glob",
+    "inquirer",
+    "semver",
+    "yaml",
+    "yargs"
+  ],
   "//": "~~ Generated by projen. To modify, edit .projenrc.js and run \"npx projen\"."
 }
