+++ src/node-project.ts
@@ -1,132 +1,34 @@
-import * as path from 'path';
-import * as fs from 'fs-extra';
-import * as semver from 'semver';
 import { PROJEN_RC, PROJEN_VERSION } from './common';
-import { DependencyType } from './deps/model';
 import { GithubWorkflow } from './github';
 import { DependabotOptions } from './github/dependabot';
 import { Mergify, MergifyOptions } from './github/mergify';
 import { IgnoreFile } from './ignore-file';
 import { Jest, JestOptions } from './jest';
-import { JsonFile } from './json';
 import { License } from './license';
-import * as logging from './logging';
+import { NpmPackage, NpmTaskExecution, NodePackageManager, NpmPackageOptions } from './npm-package';
 import { Project, ProjectOptions } from './project';
 import { ProjenUpgrade } from './projen-upgrade';
 import { Semver } from './semver';
 import { Task, TaskCategory } from './tasks';
-import { exec, isTruthy, writeFile } from './util';
 import { Version } from './version';
 
 const PROJEN_SCRIPT = 'projen';
 
-/**
- * The node package manager to use.
- */
-export enum NodePackageManager {
+export interface NodeProjectOptions extends ProjectOptions, NpmPackageOptions {
   /**
-   * Use `yarn` as the package manager.
+   * License copyright owner.
+   *
+   * @default - defaults to the value of authorName or "" if `authorName` is undefined.
    */
-  YARN = 'yarn',
+  readonly copyrightOwner?: string;
 
   /**
-   * Use `npm` as the package manager.
+   * The copyright years to put in the LICENSE file.
+   *
+   * @default - current year
    */
-  NPM = 'npm'
-}
+  readonly copyrightPeriod?: string;
 
-export interface NodeProjectCommonOptions extends ProjectOptions {
-  /**
-   * Runtime dependencies of this module.
-   *
-   * The recommendation is to only specify the module name here (e.g.
-   * `express`). This will behave similar to `yarn add` or `npm install` in the
-   * sense that it will add the module as a dependency to your `package.json`
-   * file with the latest version (`^`). You can specify semver requirements in
-   * the same syntax passed to `npm i` or `yarn add` (e.g. `express@^2`) and
-   * this will be what you `package.json` will eventually include.
-   *
-   * @example [ 'express', 'lodash', 'foo@^2' ]
-   * @default []
-   */
-  readonly deps?: string[];
-
-  /**
-   * Build dependencies for this module. These dependencies will only be
-   * available in your build environment but will not be fetched when this
-   * module is consumed.
-   *
-   * The recommendation is to only specify the module name here (e.g.
-   * `express`). This will behave similar to `yarn add` or `npm install` in the
-   * sense that it will add the module as a dependency to your `package.json`
-   * file with the latest version (`^`). You can specify semver requirements in
-   * the same syntax passed to `npm i` or `yarn add` (e.g. `express@^2`) and
-   * this will be what you `package.json` will eventually include.
-   *
-   * @example [ 'typescript', '@types/express' ]
-   * @default []
-   */
-  readonly devDeps?: string[];
-
-  /**
-   * Peer dependencies for this module. Dependencies listed here are required to
-   * be installed (and satisfied) by the _consumer_ of this library. Using peer
-   * dependencies allows you to ensure that only a single module of a certain
-   * library exists in the `node_modules` tree of your consumers.
-   *
-   * Note that prior to npm@7, peer dependencies are _not_ automatically
-   * installed, which means that adding peer dependencies to a library will be a
-   * breaking change for your customers.
-   *
-   * Unless `peerDependencyOptions.pinnedDevDependency` is disabled (it is
-   * enabled by default), projen will automatically add a dev dependency with a
-   * pinned version for each peer dependency. This will ensure that you build &
-   * test your module against the lowest peer version required.
-   *
-   * @default []
-   */
-  readonly peerDeps?: string[];
-
-  /**
-   * List of dependencies to bundle into this module. These modules will be
-   * added both to the `dependencies` section and `peerDependencies` section of
-   * your `package.json`.
-   *
-   * The recommendation is to only specify the module name here (e.g.
-   * `express`). This will behave similar to `yarn add` or `npm install` in the
-   * sense that it will add the module as a dependency to your `package.json`
-   * file with the latest version (`^`). You can specify semver requirements in
-   * the same syntax passed to `npm i` or `yarn add` (e.g. `express@^2`) and
-   * this will be what you `package.json` will eventually include.
-   */
-  readonly bundledDeps?: string[];
-
-  /**
-   * Options for `peerDeps`.
-   */
-  readonly peerDependencyOptions?: PeerDependencyOptions;
-
-  /**
-   * Binary programs vended with your module.
-   *
-   * You can use this option to add/customize how binaries are represented in
-   * your `package.json`, but unless `autoDetectBin` is `false`, every
-   * executable file under `bin` will automatically be added to this section.
-   */
-  readonly bin?: Record<string, string>;
-
-  /**
-   * Automatically add all executables under the `bin` directory to your
-   * `package.json` file under the `bin` section.
-   *
-   * @default true
-   */
-  readonly autoDetectBin?: boolean;
-
-  /**
-   * Keywords to include in `package.json`.
-   */
-  readonly keywords?: string[];
 
   /**
    * Version of projen to install.
@@ -227,20 +129,6 @@ export interface NodeProjectCommonOptions extends ProjectOptions {
   readonly antitamper?: boolean;
 
   /**
-   * Minimum Node.js version to require via package.json `engines` (inclusive).
-   *
-   * @default - no "engines" specified
-   */
-  readonly minNodeVersion?: string;
-
-  /**
-   * Minimum node.js version to require via `engines` (inclusive).
-   *
-   * @default - no max
-   */
-  readonly maxNodeVersion?: string;
-
-  /**
    * The node version to use in GitHub workflows.
    *
    * @default - same as `minNodeVersion`
@@ -260,27 +148,6 @@ export interface NodeProjectCommonOptions extends ProjectOptions {
    * @default "registry.npmjs.org"
    */
   readonly npmRegistry?: string;
-
-  /**
-   * The Node Package Manager used to execute scripts
-   *
-   * @default NodePackageManager.YARN
-   */
-  readonly packageManager?: NodePackageManager;
-
-  /**
-   * License copyright owner.
-   *
-   * @default - defaults to the value of authorName or "" if `authorName` is undefined.
-   */
-  readonly copyrightOwner?: string;
-
-  /**
-   * The copyright years to put in the LICENSE file.
-   *
-   * @default - current year
-   */
-  readonly copyrightPeriod?: string;
 
   /**
    * Compiler artifacts output directory
@@ -341,14 +208,6 @@ export interface NodeProjectCommonOptions extends ProjectOptions {
   readonly mergifyAutoMergeLabel?: string;
 
   /**
-   * npm scripts to include. If a script has the same name as a standard script,
-   * the standard script will be overwritten.
-   *
-   * @default {}
-   */
-  readonly scripts?: { [name: string]: string };
-
-  /**
    * Periodically submits a pull request for projen upgrades (executes `yarn
    * projen:upgrade`).
    *
@@ -379,14 +238,6 @@ export interface NodeProjectCommonOptions extends ProjectOptions {
    */
   readonly projenUpgradeSchedule?: string[];
 
-  /**
-   * Allow the project to include `peerDependencies` and `bundledDependencies`.
-   * This is normally only allowed for libraries. For apps, there's no meaning
-   * for specifying these.
-   *
-   * @default true
-   */
-  readonly allowLibraryDependencies?: boolean;
 
   /**
    * Defines an .npmignore file. Normally this is only needed for libraries that
@@ -402,15 +253,6 @@ export interface NodeProjectCommonOptions extends ProjectOptions {
   readonly npmignore?: string[];
 
   /**
-   * Module entrypoint (`main` in `package.json`)
-   *
-   * Set to an empty string to not include `main` in your package.json
-   *
-   * @default "lib/index.js"
-   */
-  readonly entrypoint?: string;
-
-  /**
    * Include a GitHub pull request template.
    *
    * @default true
@@ -423,21 +265,6 @@ export interface NodeProjectCommonOptions extends ProjectOptions {
    * @default - default content
    */
   readonly pullRequestTemplateContents?: string;
-
-  /**
-   * Determines how tasks are executed when invoked as npm scripts (yarn/npm run xyz).
-   * @default NpmTaskExecution.PROJEN
-   */
-  readonly npmTaskExecution?: NpmTaskExecution;
-
-  /**
-   * The shell command to use in order to run the projen CLI.
-   *
-   * Can be used to customize in special environments.
-   *
-   * @default "npx projen"
-   */
-  readonly projenCommand?: string;
 
   /**
    * Installs a GitHub workflow which is triggered when the comment "@projen
@@ -457,104 +284,6 @@ export interface NodeProjectCommonOptions extends ProjectOptions {
    * @default "rebuild"
    */
   readonly rebuildBotCommand?: string;
-
-}
-
-export enum NpmTaskExecution {
-  /**
-   * `package.json` scripts invoke to the projen CLI.
-   *
-   * @example
-   *
-   * scripts: {
-   *   "compile": "projen compile"
-   * }
-   */
-  PROJEN = 'projen',
-
-  /**
-   * Task is implemented directly as a shell script within `package.json`.
-   *
-   * @example
-   *
-   * scripts: {
-   *   "compile": "tsc"
-   * }
-   */
-  SHELL = 'shell'
-}
-
-export interface NodeProjectOptions extends NodeProjectCommonOptions {
-  /**
-   * This is the name of your package. It gets used in URLs, as an argument on the command line,
-   * and as the directory name inside node_modules.
-   * See https://classic.yarnpkg.com/en/docs/package-json/#toc-name
-   *
-   * @default $BASEDIR
-   */
-  readonly name: string;
-
-  /**
-   * The description is just a string that helps people understand the purpose of the package.
-   * It can be used when searching for packages in a package manager as well.
-   * See https://classic.yarnpkg.com/en/docs/package-json/#toc-description
-   */
-  readonly description?: string;
-
-  /**
-   * The repository is the location where the actual code for your package lives.
-   * See https://classic.yarnpkg.com/en/docs/package-json/#toc-repository
-   */
-  readonly repository?: string;
-
-  /**
-   * If the package.json for your package is not in the root directory (for example if it is part of a monorepo),
-   * you can specify the directory in which it lives.
-   */
-  readonly repositoryDirectory?: string;
-
-  /**
-   * Author's name
-   */
-  readonly authorName?: string;
-
-  /**
-   * Author's e-mail
-   */
-  readonly authorEmail?: string;
-
-  /**
-   * Author's URL / Website
-   */
-  readonly authorUrl?: string;
-
-  /**
-   * Author's Organization
-   */
-  readonly authorOrganization?: boolean;
-
-  /**
-   * Package's Homepage / Website
-   */
-  readonly homepage?: string;
-
-  /**
-   * License's SPDX identifier.
-   * See https://github.com/projen/projen/tree/master/license-text for a list of supported licenses.
-   */
-  readonly license?: string;
-
-  /**
- * Indicates if a license should be added.
- *
- * @default true
- */
-  readonly licensed?: boolean;
-
-  /**
-   * Package's Stability
-   */
-  readonly stability?: string;
 
   /**
    * Additional entries to .gitignore
@@ -593,11 +322,19 @@ export enum AutoRelease {
  * Node.js project
  */
 export class NodeProject extends Project {
+  public readonly npmPackage: NpmPackage;
   public readonly npmignore?: IgnoreFile;
   public readonly mergify?: Mergify;
-  public readonly manifest: any;
-  public readonly allowLibraryDependencies: boolean;
-  public readonly entrypoint: string;
+
+  /**
+   * @deprecated use `npmPackage.allowLibraryDependencies`
+   */
+  public get allowLibraryDependencies(): boolean { return this.npmPackage.allowLibraryDependencies; }
+
+  /**
+   * @deprecated use `npmPackage.entrypoint`
+   */
+  public get entrypoint(): string { return this.npmPackage.entrypoint; }
 
   /**
    * Compiles the code. By default for node.js projects this task is empty.
@@ -619,10 +356,6 @@ export class NodeProject extends Project {
    */
   public readonly buildTask: Task;
 
-  private readonly peerDependencyOptions: PeerDependencyOptions;
-  private readonly scripts: Record<string, string[]>;
-  private readonly bin: Record<string, string> = { };
-  private readonly keywords: Set<string>;
   private readonly _version: Version;
 
   /**
@@ -637,8 +370,8 @@ export class NodeProject extends Project {
   protected readonly releaseWorkflow?: GithubWorkflow;
   protected readonly releaseWorkflowJobId?: string;
 
-  public readonly minNodeVersion?: string;
-  public readonly maxNodeVersion?: string;
+  // public get minNodeVersion(): string | undefined { return this.npmPackage.minNodeVersion; }
+  // public get maxNodeVersion(): string | undefined { return this.npmPackage.maxNodeVersion; }
 
   private readonly nodeVersion?: string;
 
@@ -653,8 +386,10 @@ export class NodeProject extends Project {
 
   /**
    * The package manager to use.
+   *
+   * @deprecated use `npmPackage.packageManager`
    */
-  protected readonly packageManager: NodePackageManager;
+  public get packageManager(): NodePackageManager { return this.npmPackage.packageManager; }
 
   /**
    * The command to use to run scripts (e.g. `yarn run` or `npm run` depends on the package manager).
@@ -673,24 +408,28 @@ export class NodeProject extends Project {
 
   /**
    * Determines how tasks are executed when invoked as npm scripts (yarn/npm run xyz).
+   *
+   * @deprecated use `npmPackage.npmTaskExecution`
    */
-  public readonly npmTaskExecution: NpmTaskExecution;
+  public get npmTaskExecution(): NpmTaskExecution { return this.npmPackage.npmTaskExecution; }
 
   /**
    * The command to use in order to run the projen CLI.
    */
-  public readonly projenCommand: string;
+  public get projenCommand(): string { return this.npmPackage.projenCommand; }
 
-  private _renderedDeps?: NpmDependencies;
+  /**
+   * @deprecated use `npmPackage.addField(x, y)`
+   */
+  public get manifest() {
+    return this.npmPackage.manifest;
+  }
 
   constructor(options: NodeProjectOptions) {
     super(options);
 
-    this.scripts = {};
-    this.allowLibraryDependencies = options.allowLibraryDependencies ?? true;
-    this.peerDependencyOptions = options.peerDependencyOptions ?? {};
-    this.packageManager = options.packageManager ?? NodePackageManager.YARN;
-    this.npmTaskExecution = options.npmTaskExecution ?? NpmTaskExecution.PROJEN;
+    this.npmPackage = new NpmPackage(this, options);
+
     this.runScriptCommand = (() => {
       switch (this.packageManager) {
         case NodePackageManager.NPM: return 'npm run';
@@ -698,19 +437,8 @@ export class NodeProject extends Project {
         default: throw new Error(`unexpected package manager ${this.packageManager}`);
       }
     })();
-    this.projenCommand = options.projenCommand ?? 'npx projen';
 
-    this.processDeps(options);
-
-    // node version
-
-    this.minNodeVersion = options.minNodeVersion;
-    this.maxNodeVersion = options.maxNodeVersion;
-    this.nodeVersion = options.workflowNodeVersion ?? this.minNodeVersion;
-
-
-    this.keywords = new Set();
-    this.addKeywords(...options.keywords ?? []);
+    this.nodeVersion = options.workflowNodeVersion ?? this.npmPackage.minNodeVersion;
 
     // add PATH for all tasks which includes the project's npm .bin list
     this.tasks.addEnvironment('PATH', '$(npx -c \'node -e "console.log(process.env.PATH)"\')');
@@ -737,77 +465,13 @@ export class NodeProject extends Project {
       category: TaskCategory.BUILD,
     });
 
-    let nodeVersion = '';
-
-    if (this.minNodeVersion) {
-      nodeVersion += `>= ${this.minNodeVersion}`;
-    }
-
-    if (this.maxNodeVersion) {
-      nodeVersion += ` <= ${this.maxNodeVersion}`;
-    }
+    this.addLicense(options);
 
     this.npmDistTag = options.npmDistTag ?? 'latest';
     this.npmRegistry = options.npmRegistry ?? 'registry.npmjs.org';
 
-    const renderScripts = () => {
-      const result: any = {};
-      for (const [name, commands] of Object.entries(this.scripts)) {
-        const cmds = commands.length > 0 ? commands : ['echo "n/a"'];
-        result[name] = cmds.join(' && ');
-      }
-      for (const task of this.tasks.all) {
-        result[task.name] = this.npmScriptForTask(task);
-      }
-
-      return result;
-    };
-
-    let author;
-
-    if (options.authorName) {
-      author = {
-        name: options.authorName,
-        email: options.authorEmail,
-        url: options.authorUrl,
-        organization: options.authorOrganization ?? false,
-      };
-    } else {
-      if (options.authorEmail || options.authorUrl || options.authorOrganization !== undefined) {
-        throw new Error('"authorName" is required if specifying "authorEmail" or "authorUrl"');
-      }
-    }
-
     this.testdir = options.testdir ?? 'test';
 
-    this.manifest = {
-      name: options.name,
-      description: options.description,
-      repository: !options.repository ? undefined : {
-        type: 'git',
-        url: options.repository,
-        directory: options.repositoryDirectory,
-      },
-      bin: this.bin,
-      scripts: renderScripts,
-      author: author,
-      homepage: options.homepage,
-      devDependencies: {},
-      peerDependencies: {},
-      dependencies: {},
-      bundledDependencies: [],
-      keywords: () => Array.from(this.keywords).sort(),
-      engines: nodeVersion !== '' ? { node: nodeVersion } : undefined,
-    };
-
-    this.entrypoint = options.entrypoint ?? 'lib/index.js';
-    this.manifest.main = this.entrypoint !== '' ? this.entrypoint : undefined;
-
-    new JsonFile(this, 'package.json', {
-      obj: this.manifest,
-      marker: true,
-      readonly: false, // we want "yarn add" to work and we have anti-tamper
-    });
 
     if (options.npmignoreEnabled ?? true) {
       this.npmignore = new IgnoreFile(this, '.npmignore');
@@ -831,26 +495,13 @@ export class NodeProject extends Project {
       }
     }
 
-    // set license and produce license file
-    if (options.licensed ?? true) {
-      const license = options.license ?? 'Apache-2.0';
-      this.manifest.license = license;
 
-      new License(this, license, {
-        copyrightOwner: options.copyrightOwner ?? options.authorName,
-        copyrightPeriod: options.copyrightPeriod,
-      });
-    } else {
-      this.manifest.license = 'UNLICENSED';
-    }
-
-    this.setScript(PROJEN_SCRIPT, this.projenCommand);
-    this.setScript('start', `${this.projenCommand} start`);
+    this.setScript(PROJEN_SCRIPT, this.npmPackage.projenCommand);
+    this.setScript('start', `${this.npmPackage.projenCommand} start`);
 
     this.npmignore?.exclude(`/${PROJEN_RC}`);
     this.gitignore.include(`/${PROJEN_RC}`);
 
-    this.addBins(options.bin ?? { });
 
     const projen = options.projenDevDependency ?? true;
     if (projen) {
@@ -862,7 +513,7 @@ export class NodeProject extends Project {
 
     // version is read from a committed file called version.json which is how we bump
     this._version = new Version(this, { releaseBranch: defaultReleaseBranch });
-    this.manifest.version = (outdir: string) => this._version.resolveVersion(outdir);
+    this.npmPackage.addVersion(this._version.currentVersion);
 
     // indicate if we have anti-tamper configured in our workflows. used by e.g. Jest
     // to decide if we can always run with --updateSnapshot
@@ -973,21 +624,6 @@ export class NodeProject extends Project {
       }
     }
 
-    // automatically add all executable files under "bin"
-    if (options.autoDetectBin ?? true) {
-      const bindir = 'bin';
-
-      if (fs.existsSync(bindir)) {
-        for (const file of fs.readdirSync(bindir)) {
-          try {
-            fs.accessSync(path.join(bindir, file), fs.constants.X_OK);
-            this.bin[file] = path.join(bindir, file).replace(/\\/g, '/');
-          } catch (e) {
-            // not executable, skip
-          }
-        }
-      }
-    }
 
     let autoMergeLabel;
 
@@ -1048,11 +684,6 @@ export class NodeProject extends Project {
       labels: (projenAutoMerge && autoMergeLabel) ? [autoMergeLabel] : [],
     });
 
-    // override any scripts from options (if specified)
-    for (const [cmdname, shell] of Object.entries(options.scripts ?? {})) {
-      this.addTask(cmdname, { exec: shell });
-    }
-
     if (options.pullRequestTemplate ?? true) {
       this.github?.addPullRequestTemplate(...options.pullRequestTemplateContents ?? []);
     }
@@ -1064,9 +695,7 @@ export class NodeProject extends Project {
   }
 
   public addBins(bins: Record<string, string>) {
-    for (const [k, v] of Object.entries(bins)) {
-      this.bin[k] = v;
-    }
+    this.npmPackage.addBin(bins);
   }
 
   /**
@@ -1076,7 +705,7 @@ export class NodeProject extends Project {
    * @param command The command to execute
    */
   public setScript(name: string, command: string) {
-    this.scripts[name] = [command];
+    this.npmPackage.setScript(name, command);
   }
 
   /**
@@ -1084,16 +713,15 @@ export class NodeProject extends Project {
    * @param name The name of the script.
    */
   public removeScript(name: string) {
-    delete this.scripts[name];
+    this.npmPackage.removeScript(name);
   }
-
 
   /**
    * Indicates if a script by the name name is defined.
    * @param name The name of the script
    */
   public hasScript(name: string) {
-    return name in this.scripts;
+    return this.npmPackage.hasScript(name);
   }
 
   /**
@@ -1132,7 +760,7 @@ export class NodeProject extends Project {
    */
   public addFields(fields: { [name: string]: any }) {
     for (const [name, value] of Object.entries(fields)) {
-      this.manifest[name] = value;
+      this.npmPackage.addField(name, value);
     }
   }
 
@@ -1141,9 +769,7 @@ export class NodeProject extends Project {
    * @param keywords The keywords to add
    */
   public addKeywords(...keywords: string[]) {
-    for (const k of keywords) {
-      this.keywords.add(k);
-    }
+    this.npmPackage.addKeywords(...keywords);
   }
 
   public get installWorkflowSteps(): any[] {
@@ -1158,13 +784,13 @@ export class NodeProject extends Project {
 
     install.push({
       name: 'Install dependencies',
-      run: this.renderInstallCommand(true),
+      run: this.npmPackage.installCommand,
     });
 
     // run "projen"
     install.push({
       name: 'Synthesize project files',
-      run: this.projenCommand,
+      run: this.npmPackage.projenCommand,
     });
 
     return install;
@@ -1180,9 +806,7 @@ export class NodeProject extends Project {
    * `module@^7`.
    */
   public addDeps(...deps: string[]) {
-    for (const dep of deps) {
-      this.deps.addDependency(dep, DependencyType.RUNTIME);
-    }
+    return this.npmPackage.addDeps(...deps);
   }
 
   /**
@@ -1195,10 +819,7 @@ export class NodeProject extends Project {
    * `module@^7`.
    */
   public addDevDeps(...deps: string[]) {
-    for (const dep of deps) {
-      this.deps.addDependency(dep, DependencyType.BUILD);
-
-    }
+    return this.npmPackage.addDevDeps(...deps);
   }
 
   /**
@@ -1215,13 +836,7 @@ export class NodeProject extends Project {
    * `module@^7`.
    */
   public addPeerDeps(...deps: string[]) {
-    if (Object.keys(deps).length && !this.allowLibraryDependencies) {
-      throw new Error(`cannot add peer dependencies to an APP project: ${Object.keys(deps).join(',')}`);
-    }
-
-    for (const dep of deps) {
-      this.deps.addDependency(dep, DependencyType.PEER);
-    }
+    return this.npmPackage.addPeerDeps(...deps);
   }
 
   /**
@@ -1237,226 +852,19 @@ export class NodeProject extends Project {
    * `module@^7`.
    */
   public addBundledDeps(...deps: string[]) {
-    if (deps.length && !this.allowLibraryDependencies) {
-      throw new Error(`cannot add bundled dependencies to an APP project: ${deps.join(',')}`);
-    }
-
-    for (const dep of deps) {
-      this.deps.addDependency(dep, DependencyType.BUNDLED);
-    }
+    return this.npmPackage.addBundledDeps(...deps);
   }
 
-  private processDeps(options: NodeProjectCommonOptions) {
-    const deprecate = (key: string, alt: string) => {
-      if (Object.keys((options as any)[key] ?? {}).length) {
-        throw new Error(`The option "${key}" is no longer supported, use "${alt}" instead (see API docs)`);
-      }
-    };
-
-    deprecate('dependencies', 'deps');
-    deprecate('peerDependencies', 'peerDeps');
-    deprecate('devDependencies', 'devDeps');
-    deprecate('bundledDependencies', 'bundledDeps');
-
-    this.addDeps(...options.deps ?? []);
-    this.addDevDeps(...options.devDeps ?? []);
-    this.addPeerDeps(...options.peerDeps ?? []);
-    this.addBundledDeps(...options.bundledDeps ?? []);
-  }
-
-  public preSynthesize() {
-    this._renderedDeps = this.renderDependencies();
-  }
-
-  public postSynthesize() {
-    super.postSynthesize();
-
-    const outdir = this.outdir;
-
-    // now we run `yarn install`, but before we do that, remove the
-    // `node_modules/projen` symlink so that yarn won't hate us.
-    const projenModule = path.resolve('node_modules', 'projen');
-    try {
-      if (fs.lstatSync(projenModule).isSymbolicLink()) {
-        fs.unlinkSync(projenModule);
-      }
-    } catch (e) { }
-
-    exec(this.renderInstallCommand(isTruthy(process.env.CI)), { cwd: outdir });
-
-    this.resolveDepsAndWritePackageJson(outdir);
-  }
-
-  private renderInstallCommand(frozen: boolean) {
-    switch (this.packageManager) {
-      case NodePackageManager.YARN:
-        return [
-          'yarn install',
-          '--check-files', // ensure all modules exist (especially projen which was just removed).
-          ...frozen ? ['--frozen-lockfile'] : [],
-        ].join(' ');
-
-      case NodePackageManager.NPM:
-        return frozen
-          ? 'npm ci'
-          : 'npm install';
-
-      default:
-        throw new Error(`unexpected package manager ${this.packageManager}`);
+  private addLicense(options: NodeProjectOptions) {
+    if (this.npmPackage.license) {
+      new License(this, this.npmPackage.license, {
+        copyrightOwner: options.copyrightOwner ?? options.authorName,
+        copyrightPeriod: options.copyrightPeriod,
+      });
     }
-  }
-
-  private renderDependencies(): NpmDependencies {
-    const devDependencies: Record<string, string> = {};
-    const peerDependencies: Record<string, string> = {};
-    const dependencies: Record<string, string> = {};
-    const bundledDependencies = new Array<string>();
-
-    // synthetic dependencies: add a pinned build dependency to ensure we are
-    // testing against the minimum requirement of the peer.
-    const pinned = this.peerDependencyOptions.pinnedDevDependency ?? true;
-    if (pinned) {
-      for (const dep of this.deps.all.filter(d => d.type === DependencyType.PEER)) {
-        let req = dep.name;
-        if (dep.version) {
-          const ver = semver.minVersion(dep.version)?.version;
-          if (!ver) {
-            throw new Error(`unable to determine minimum semver for peer dependency ${dep.name}@${dep.version}`);
-          }
-
-          req += '@' + ver;
-        }
-        this.addDevDeps(req);
-      }
-    }
-
-    for (const dep of this.deps.all) {
-      const version = dep.version ?? '*';
-
-      switch (dep.type) {
-        case DependencyType.BUNDLED:
-          bundledDependencies.push(dep.name);
-
-          if (this.deps.all.find(d => d.name === dep.name && d.type === DependencyType.PEER)) {
-            throw new Error(`unable to bundle "${dep.name}". it cannot appear as a peer dependency`);
-          }
-
-          // also add as a runtime dependency
-          dependencies[dep.name] = version;
-          break;
-
-        case DependencyType.PEER:
-          peerDependencies[dep.name] = version;
-          break;
-
-        case DependencyType.RUNTIME:
-          dependencies[dep.name] = version;
-          break;
-
-        case DependencyType.TEST:
-        case DependencyType.DEVENV:
-        case DependencyType.BUILD:
-          devDependencies[dep.name] = version;
-          break;
-      }
-    }
-
-    // update the manifest we are about to save into `package.json`
-    this.manifest.devDependencies = devDependencies;
-    this.manifest.peerDependencies = peerDependencies;
-    this.manifest.dependencies = dependencies;
-    this.manifest.bundledDependencies = bundledDependencies;
-
-    // nothing further to do if package.json file does not exist
-    const root = path.join(this.outdir, 'package.json');
-    if (!fs.existsSync(root)) {
-      return { devDependencies, peerDependencies, dependencies };
-    }
-
-    const pkg = JSON.parse(fs.readFileSync(root, 'utf-8'));
-
-    const readDeps = (user: Record<string, string>, current: Record<string, string> = {}) => {
-      for (const [name, userVersion] of Object.entries(user)) {
-        const currentVersion = current[name];
-
-        // respect user version if it's not '*' or if current version is undefined
-        if (userVersion !== '*' || !currentVersion || currentVersion === '*') {
-          continue;
-        }
-
-        // memoize current version in memory so it is preserved when saving
-        user[name] = currentVersion;
-      }
-
-      // report removals
-      for (const name of Object.keys(current ?? {})) {
-        if (!user[name]) {
-          logging.verbose(`${name}: removed`);
-        }
-      }
-    };
-
-    readDeps(devDependencies, pkg.devDependencies);
-    readDeps(dependencies, pkg.dependencies);
-    readDeps(peerDependencies, pkg.peerDependencies);
-
-    return { devDependencies, dependencies, peerDependencies };
-  }
-
-  private resolveDepsAndWritePackageJson(outdir: string) {
-    const root = path.join(outdir, 'package.json');
-    const pkg = JSON.parse(fs.readFileSync(root, 'utf-8'));
-
-    const resolveDeps = (current: { [name: string]: string }, user: Record<string, string>) => {
-      const result: Record<string, string> = {};
-
-      for (const [name, currentDefinition] of Object.entries(user)) {
-        // find actual version from node_modules
-        let desiredVersion = currentDefinition;
-
-        if (currentDefinition === '*') {
-          try {
-            const modulePath = require.resolve(`${name}/package.json`, { paths: [outdir] });
-            const module = JSON.parse(fs.readFileSync(modulePath, 'utf-8'));
-            desiredVersion = `^${module.version}`;
-          } catch (e) { }
-
-          if (!desiredVersion) {
-            logging.warn(`unable to resolve version for ${name} from installed modules`);
-            continue;
-          }
-        }
-
-        if (currentDefinition !== desiredVersion) {
-          logging.verbose(`${name}: ${currentDefinition} => ${desiredVersion}`);
-        }
-
-        result[name] = desiredVersion;
-      }
-
-      // print removed packages
-      for (const name of Object.keys(current)) {
-        if (!result[name]) {
-          logging.verbose(`${name} removed`);
-        }
-      }
-
-      return sorted(result)();
-    };
-
-    const rendered = this._renderedDeps;
-    if (!rendered) {
-      throw new Error('assertion failed');
-    }
-    pkg.dependencies = resolveDeps(pkg.dependencies, rendered.dependencies);
-    pkg.devDependencies = resolveDeps(pkg.devDependencies, rendered.devDependencies);
-    pkg.peerDependencies = resolveDeps(pkg.peerDependencies, rendered.peerDependencies);
-
-    writeFile(root, JSON.stringify(pkg, undefined, 2));
   }
 
   private addDefaultGitIgnore() {
-
     this.gitignore.exclude(
       '# Logs',
       'logs',
@@ -1645,20 +1053,11 @@ export class NodeProject extends Project {
    * @param task The task for which the command is required
    */
   public runTaskCommand(task: Task) {
-    switch (this.npmTaskExecution) {
-      case NpmTaskExecution.PROJEN: return `${this.projenCommand} ${task.name}`;
+    switch (this.npmPackage.npmTaskExecution) {
+      case NpmTaskExecution.PROJEN: return `${this.npmPackage.projenCommand} ${task.name}`;
       case NpmTaskExecution.SHELL: return `${this.runScriptCommand} ${task.name}`;
       default:
-        throw new Error(`invalid npmTaskExecution mode: ${this.npmTaskExecution}`);
-    }
-  }
-
-  private npmScriptForTask(task: Task) {
-    switch (this.npmTaskExecution) {
-      case NpmTaskExecution.PROJEN: return `${this.projenCommand} ${task.name}`;
-      case NpmTaskExecution.SHELL: return task.toShellCommand();
-      default:
-        throw new Error(`invalid npmTaskExecution mode: ${this.npmTaskExecution}`);
+        throw new Error(`invalid npmTaskExecution mode: ${this.npmPackage.npmTaskExecution}`);
     }
   }
 
@@ -1721,14 +1120,6 @@ interface BuildWorkflow {
   readonly buildJobId: string;
 }
 
-export interface PeerDependencyOptions {
-  /**
-   * Automatically add a pinned dev dependency.
-   * @default true
-   */
-  readonly pinnedDevDependency?: boolean;
-}
-
 interface NodeBuildWorkflowOptions {
   /**
    * @default - default image
@@ -1787,26 +1178,4 @@ interface NodeBuildWorkflowOptions {
 export interface NodeWorkflowSteps {
   readonly antitamper: any[];
   readonly install: any[];
-}
-
-interface NpmDependencies {
-  readonly dependencies: Record<string, string>;
-  readonly devDependencies: Record<string, string>;
-  readonly peerDependencies: Record<string, string>;
-}
-
-function sorted<T>(toSort: T) {
-  return () => {
-    if (Array.isArray(toSort)) {
-      return (toSort as unknown[]).sort();
-    } else if (toSort != null && typeof toSort === 'object') {
-      const result: Record<string, unknown> = {};
-      for (const [key, value] of Object.entries(toSort).sort(([l], [r]) => l.localeCompare(r))) {
-        result[key] = value;
-      }
-      return result as T;
-    } else {
-      return toSort;
-    }
-  };
 }
