+++ src/typescript.ts
@@ -158,7 +158,8 @@ export class TypeScriptProject extends NodeProject {
     }
 
     if (options.entrypointTypes || this.entrypoint !== '') {
-      this.manifest.types = options.entrypointTypes ?? `${path.join(path.dirname(this.entrypoint), path.basename(this.entrypoint, '.js')).replace(/\\/g, '/')}.d.ts`;
+      const entrypointTypes = options.entrypointTypes ?? `${path.join(path.dirname(this.entrypoint), path.basename(this.entrypoint, '.js')).replace(/\\/g, '/')}.d.ts`;
+      this.npmPackage.addField('types', entrypointTypes);
     }
 
     const compilerOptionDefaults: TypeScriptCompilerOptions = {
@@ -263,7 +264,7 @@ export class TypeScriptProject extends NodeProject {
 
     this.addDevDeps(
       `typescript@${options.typescriptVersion ?? '^3.9.5'}`,
-      `@types/node@^${this.minNodeVersion ?? '10.17.0'}`, // install the minimum version to ensure compatibility
+      `@types/node@^${this.npmPackage.minNodeVersion ?? '10.17.0'}`, // install the minimum version to ensure compatibility
     );
 
     // generate sample code in `src` and `lib` if these directories are empty or non-existent.
