+++ src/util.ts
@@ -167,3 +167,24 @@ export function deepMerge(...objects: Array<Obj<any> | undefined>) {
 export function dedupArray<T>(array: T[]): T[] {
   return array.filter((val, idx) => array.indexOf(val) === idx);
 }
+
+/**
+ * Returns a sorted version of `x` or `undefined` if it is an empty array or object.
+ */
+export function sorted<T>(x: T) {
+  if (x == null) { return undefined; }
+  if (Array.isArray(x)) {
+    if (x.length === 0) { return undefined; }
+    return (x as unknown[]).sort();
+  } else if (typeof x === 'object') {
+    if (Object.keys(x).length === 0) { return undefined; }
+    const result: Record<string, unknown> = {};
+    for (const [key, value] of Object.entries(x).sort(([l], [r]) => l.localeCompare(r))) {
+      result[key] = value;
+    }
+    return result as T;
+  } else {
+    return x;
+  }
+}
+
