+++ test/npm-package.test.ts
@@ -1,0 +1,21 @@
+import { NpmPackage } from '../src';
+import { synthSnapshot, TestProject } from './util';
+
+test('minimal', () => {
+  // GIVEN
+  const project = new TestProject();
+
+  // WHEN
+  new NpmPackage(project, {
+    name: 'my-project',
+  });
+
+  // THEN
+  const pkg = synthSnapshot(project)['package.json'];
+  expect(pkg).toStrictEqual({
+    'name': 'my-project',
+    'license': 'Apache-2.0',
+    'main': 'lib/index.js',
+    '//': '~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".',
+  });
+});
